Replace DnsTransaction server selection with DnsServerIterator
The server selection in DnsTransaction does not correct return which
server to attempt when multiple transactions are racing.

Replace this method with a DnsServerIterator , which is scoped per
transaction and not subject to racing, to return the correct server.

The DnsServerIterator interface is implemented by Secure DNS (HTTP)
and classic DNS (UDP) flavors.

Bug: 1045507
Change-Id: Ia3eb52cfb4c384184251825751d6140fda91a4ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2095453
Commit-Queue: Steven Bingler <bingler@chromium.org>
Reviewed-by: Eric Orth <ericorth@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#749969}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 382d7d4f27e0fe4997f8182e9e205b6dbbef7b7a