Make TransportSecurityState::ShouldRequireCT a boolean
Currently, ShouldRequireCTForTesting is implemented as a tristate,
with values of false, true, and nullptr. true and nullptr work as
expected, while false causes CT to be required in some cases where
it is otherwise not required (such as official builds, for locally
trusted certs).

This simplifies the implementation by making it bi-state, and
renames the method and enums to better reflect that true is forcing
CT on absolutely, while false is... not forcing CT on absolutely.

Bug: 1064391
Change-Id: I1f85a70d32e5c9a0772d877f8217009e6ccddf49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2118577
Reviewed-by: David Benjamin <davidben@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Commit-Queue: Ryan Sleevi <rsleevi@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#753364}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9edc2e6873beff229c44bf7ef4ba726818f9b4ef