Avoid checking parental controls on startup
Start a 2 second timer on SystemNetworkContextManager creation.  On
fire, checks for parental controls and updates the network service DNS
config if DoH needs to be disabled. Should be more performant to check
after the delay because the actual underlying API calls with performance
issues are already triggered off-thread in Chrome main(), so the check
is only a performance issue if called too soon after startup.

On all config retrievals, skips checking parental controls if it hasn't
been checked yet and the otherwise-configured mode is AUTOMATIC. In this
mode, DoH activation probes don't even start until 5 seconds after
NetworkService creation, so delaying 2 seconds before disabling DoH
should have no practical effect.  If the otherwise-configured mode is
OFF, it doesn't ever need to be checked, and if SECURE, it's acceptable
to take the performance hit to always wait for the check.

For UI retrieval of config, set a flag to force checking parental
controls immediately to ensure the UI shows the correct state. If the
check hadn't been made yet, this triggers an update to NetworkService to
ensure the UI state matches the service state.  Made some simple changes
to pull the service update calls into GetStubResolverConfig() to
simplify its ability to perform the service update.

Also converted the type for retrieving DoH servers from the mojo type to
DnsOverHttpsServerConfig now that the mojo update is handled internally
to GetStubResolverConfig(), and moved the type to doh_provider_list.h to
be more widely shareable.  Little bit of code churn, but nothing
significant.

Bug: 1064346
Change-Id: I6505bd363d6c1fb88b255dd8c1dd1973b721bf97
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2130867
Reviewed-by: Esmael Elmoslimany <aee@chromium.org>
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Eric Orth <ericorth@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#756421}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c63cf50bc13169e3bf01f6ae2bdb71a881a31a6f