Add NetworkIsolationKey field to ReportingReport
This adds a NetworkIsolationKey field to ReportingReport, which
represents the report to be sent via Reporting API to some previously
configured endpoint.

This NetworkIsolationKey will eventually be filled in with the
NetworkIsolationKey of the request for which the report was generated.

For now this field is always empty, except in tests.

Bug: 993805
Change-Id: Ie957d4600007f4a46c4fc32b3381d55f518b1a42
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2137924
Commit-Queue: Lily Chen <chlily@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#758027}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0938346c775555be9655f53516d69eb6b2aa80e9