Fix EV PageInfo check
Before we required Organization, Country, and Locality to be set, but EV
certificates require Organization, Country, and at least one of Locality
OR State. This updates the check in PageInfoBubbleView to only check the
fields that are actually used in the UI, and adds a regression test (and
certificate) exercising this case.

Fixed: 1069113
Change-Id: I5a5eff339eb875fa4d36c61f0e5ffc1d4ba95ce3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2142714
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Commit-Queue: Christopher Thompson <cthomp@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#758296}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 973da38fd34f44571f1f827345be3364f468807a