Remove URLRequestHttpGet::GetStatus().
It was only used by a pair of DCHECKs that can check if the job is
"done" instead. Done includes cancellation, but also other completion
reasons, which is actually a better thing to check at the old callsites.

This is part of an effort to remove URLRequestStatus.

Bug: 651119
Change-Id: I99454888368c7b5aa320153e1911dfa3230923f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2164690
Reviewed-by: Eric Roman <eroman@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#763001}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f47e120b52e33d276b88754398e6b17e2c3715ac