SimpleCache: Remove EntryTrailerSize histogram.
This CL removes the SimpleCache.*.EntryTrailerSize histogram.  It produces
a very large quantity of data and we are not currently looking at the
results.  If we run an experiment to launch speculative prefetching in
the future we could add it back.

Bug: 1000088
Change-Id: I55ac81b6774d32431dfe2259ba1fa2cb597b9fcf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2219073
Reviewed-by: Maksim Orlovich <morlovich@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Commit-Queue: Ben Kelly <wanderview@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#772854}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8d65b3315c5dbb716e915d52f1eaed71e8775d08