Rework TransportSecurityPersister's data format.
In particular, separate out the HSTS and ExpectCT data tables
(addressing an old TODO) and switch from a dictionary of dictionaries
to a list of dictionaries, so that ExpectCT entries can be keyed on
NetworkIsolationKey in addition to domain hash in a followup CL.

The current plan is not to partition HSTS information by
NetworkIsolationKey, but this CL also switches it to using a list, to
be similar to the ExpectCT table.

This CL also adds a version number to the saved data format, and adds
an empty NetworkIsolationKey field to ExpectCT entries, so that entries
won't be lost when that field is added (though entries may be lost
when switching from an empty NIK to actually applying NIKs, which is
an issue for another day).

Bug: 969893
Change-Id: If723386b7ebc447ae78cb393b6b9d6719441759d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2216040
Commit-Queue: Matt Menke <mmenke@chromium.org>
Reviewed-by: Nick Harper <nharper@chromium.org>
Reviewed-by: Emily Stark <estark@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#772952}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 911e448db9023cccd3f6063c57c1e4d46a48b702