Remove Net.SSLRSAKeyUsage.KnownRoot and extend UnknownRoot
The KnownRoot histogram is no longer needed because we now enforce it
unconditionally. Extend the unknown root histogram for a year as we have
not yet enforced it there.

(This doesn't address all the histgrams on the bug. I'll deal with those
in separate CLs.)

Bug: 1089024
Change-Id: I75aad10136eeb7ec7d74ca0bda70bac23cf13c6f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2225378
Reviewed-by: Matt Mueller <mattm@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Commit-Queue: David Benjamin <davidben@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#773817}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 728ad079d8e52ada4e321fb4f53713e4f0588072