Bring URLRequestJob::ComputeReferrerForRequest in sync with //content
Currently, we have two places (outside of Blink... three in total, but
that's for another day) where we compute requests' referrers, nominally
according to [1, Algorithm 8.1].

This change makes a couple alterations to
URLRequestJob::ComputeReferrerForPolicy (one of the places) to bring it
in line with the behavior in content::Referrer::SanitizeForRequest
(another). With this change, ComputeReferrerForPolicy will:
- truncate requests with referrers longer than 4096 bytes to the
initiating origin;
- respect the CapReferrerToOriginOnCrossOrigin base::Feature, which
adds an optional layer of referrer truncation on top of whatever's
specified by the operative referrer policy; and
- use GURL::GetAsReferrer to strip the referrer before returning it
[1, Algorithm 8.4].

One detail worth mentioning is the behavior change concerning requests
with referrers with !GURL::SchemeIsValidForReferrer. GURL::GetAsReferrer
actually does two things: first, it strips the URL for use as a referrer
in the sense of [1, Algorithm 8.4]; second, it returns GURL() if the
URL does not satisfy GURL::SchemeIsValidForReferrer. Previously, since
URLRequestJob::ComputeReferrerForPolicy wasn't calling GetAsReferrer, it
let URLs through with !SchemeIsValidForReferrer. After the change, it
will now return GURL() for these URLs, which seems like the desired
behavior given that it's how content::Referrer::SanitizeForRequest and
blink::SecurityPolicy::GenerateReferrer behave.

This change is a precursor to getting rid of
content::Referrer::SanitizeForRequest (more specifically, removing all
of its business logic and just delegating to
URLRequestJob::ComputeReferrerForPolicy). There is one notable way in
which it does *not* bring URLRequestJob::ComputeReferrerForPolicy
directly in line with content::Referrer::SanitizeForRequest:
content::Referrer::SanitizeForRequest currently returns an empty
referrer for requests with *destinations* of non-HTTP(S) scheme, but
this behavior is not mandated by the standard and disagrees
with the logic in both URLRequestJob and blink::SecurityPolicy; the 2014
CL implementing it [2] does not contain a specific motivation for the
behavior.

[1]:
https://w3c.github.io/webappsec-referrer-policy/#parse-referrer-policy-from-header
[2]: https://codereview.chromium.org/438283002

Bug: 1094526
Change-Id: I80c6cd51f56a220e4e3f78fa32d7e8faccd09a85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2243992
Reviewed-by: Dominic Farolino <dom@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Commit-Queue: Avi Drissman <avi@chromium.org>
Auto-Submit: David Van Cleve <davidvc@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#781922}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 09c1e888854f4ca8bf35de6b5e67fead82274ef7