[net] Introduce basic connectivity monitoring
This adds a new ConnectivityMonitor object to NetworkQualityEstimator.

ConnectivityMonitor tracks all pending network requests throughout
their lifetime while the host device is connected to a Wi-Fi network.
If a long period of inactivity (configurable, default 2500 ms) is
detected across all requests, ConnectivityMonitor takes note of this
as a potential connectivity issue.

If the new ReportPoorConnectivity feature is enabled, this issue is
reported immediately to the OS (Android only, for now) as a network
failure.

ConnectivityMonitor also watches for network changes. If it suspects
a connectivity issue and the network switches, we record a new
NQE.ConnectivityMonitor.TimeToSwitchNetworks metric as the time delta
between the original failure detection (regardless of whether it was
reported to the OS) and the actual network change.

This CL serves as the basis for an experiment to understand how OS
connectivity reporting in Chrome may impact network failure
recovery.

Bug: 1079380
Change-Id: Ieecb2bf1be1cc9cb17496b5493204e4f0e3c0ee6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2275340
Commit-Queue: Ken Rockot <rockot@google.com>
Reviewed-by: Zhongyi Shi <zhongyi@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#786166}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 16610e4d3d534e6bdde1d7f38a6231f9e85d9684