Reenable disable_draft_25
This flag was verified last week, and was then set
`chrome_value: false` because it was breaking tests.
After digging further, I realized that the failing
test was incorrect: it would create a dispatcher with
no versions, which triggers a DCHECK. This used to pass
because draft-25 was a version that could support
WebTransport but wasn't part of the origin trial
versions. I don't think this test adds much value,
because if there are no versions on the server we don't
really care about validating that everything fails. So
this CL removes that test, and flips the flag back.

Change-Id: I80bc8bd3ee76238311d3ac21b034b5a526e6202a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2297964
Auto-Submit: David Schinazi <dschinazi@chromium.org>
Reviewed-by: Nick Harper <nharper@chromium.org>
Commit-Queue: David Schinazi <dschinazi@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#788851}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 2cbd8c62c05fb7addb56ef9ed16d904d89957aa7