Cleanup dead code in DnsSocketPool
Remove everything related to the "default" socket pool that is
completely unused.  Also simplify the DnsSocketPool interface for the
remaining much simpler usage of the class.

Also delete dns_session_unittest.cc since it was really only testing
that allocated sockets were properly freed in the pool, but that no
longer matters and it wasn't worth maintaining the mockability and hooks
in the pool needed for the test.

Bug: 1116579
Change-Id: Ib322bffb653e85a9c33a4b3835f4067fd8d29530
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2357615
Commit-Queue: Eric Orth <ericorth@chromium.org>
Reviewed-by: Dan McArdle <dmcardle@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800248}
GitOrigin-RevId: 990a74a91ec1b518414b76485e8a84b72cf8e2cf