Fix the network bug triage queries when there are multiple components.
Previously, our query would select all of Internals>Network>* and then
subtract bugs with one of the out-of-scope components. The problem is
this filters out bugs that contain both an in-scope and an out-of-scope
bug component.

The new queries use an additive approach that names each of the
in-scope components directly.

NOTRY=true

Change-Id: I8526c49a38d45d41caa4d53b3c6103d281e7f4a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2367983
Reviewed-by: Eric Roman <eroman@chromium.org>
Reviewed-by: Zhongyi Shi <zhongyi@chromium.org>
Commit-Queue: Zhongyi Shi <zhongyi@chromium.org>
Auto-Submit: Eric Roman <eroman@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800644}
GitOrigin-RevId: a92a90a4402df9d20838d05df367b9790ce4fcb0