Re-land "Allow use of Abseil in //net/third_party/quiche".
This is a re-land
https://chromium-review.googlesource.com/c/chromium/src/+/2302721, but
without visibility changes (which are no longer needed).

Last time, this broke one of the Windows bots.  I appear to be unable
to even run code on those bots; however, since the last attempt, Abseil
is now linked into //base (which //net depends on), so presumably this
shouldn't break.

Change-Id: I3a3639b4406bc41e2decd116c76d4373cb5f7a1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2349689
Reviewed-by: Zhongyi Shi <zhongyi@chromium.org>
Reviewed-by: David Schinazi <dschinazi@chromium.org>
Commit-Queue: Victor Vasiliev <vasilvv@chromium.org>
Cr-Commit-Position: refs/heads/master@{#801187}
GitOrigin-RevId: 1614e0dc6df082b2054d5ce5e225a4888ce23ea3