Remove URLRequestJobManager
Back in the days when humanity was yet young, and IE roamed the Earth,
URLRequestJobManager had a global list of factory functions, one for
each scheme. Through the mighty work of the intern shalev, most
of these were converted into ProtocolHandlers, which were added to a
URLRequestJobFactoryImpl, which sat at the bottom of a stack of
URLRequestJobFactories used by various modules to to intercept requests.

However, the zombie URLRequestJobManager remained. It called into
the top URLRequestJobFactory to give it the first shot at creating a
URLRequestJob, and it still had the factory functions for HTTP[S] and
WS[S].

There were reasons for this, though I can't quite remember why -
URLRequest::IsSupportedScheme (removed a while back) was likely one of
those reasons. There may have been others.

At this point, we've removed almost all ProtocolHandlers (only FTP and
chrome (iOS only)) remain, and the intercepting URLRequestJobFactories
that sat on top of it are also gone. This CL finally removes
URLRequestJobManager and moves all of its logic down to
URLRequestJobFactoryImpl.

Fixed: 354158
Change-Id: Ic1f364bd6bf9c37ed6e139a668da1de2bb301bd0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2392995
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-by: Eric Roman <eroman@chromium.org>
Reviewed-by: Yuwei Huang <yuweih@chromium.org>
Commit-Queue: Matt Menke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/master@{#804954}
GitOrigin-RevId: 7ca3f7c491df1f3f405496e9b3b3b37e2bb065c0