Make QuicheStringPiece use absl::string_view
This changes QuicheStringPiece to be an alias of absl::string_view,
rather than base::StringPiece.  This is a first step in fully replacing
QuicheStringPiece with absl::string_view and deprecating the earlier.

This adds conversions around the code and switches some of
the text_utils_impl to use Abseil directly in order to get the code
to compile.

Bug: 1128392
Change-Id: I1034ac9b381e44668a738f6538395cf9e338d2f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2411554
Reviewed-by: David Schinazi <dschinazi@chromium.org>
Reviewed-by: Mirko Bonadei <mbonadei@chromium.org>
Commit-Queue: Victor Vasiliev <vasilvv@chromium.org>
Cr-Commit-Position: refs/heads/master@{#807234}
GitOrigin-RevId: 502ba58ffe05411aba40caef6d2e557053b5671a