Remove Apple specific socket code for getting IPv4 address.
This code is no longer needed and breaks ARM mac.

MacOS added support for the Linux-style socket option
sometime between 2008 and 2010 (see links below) and
Chrome currently only supports macOS versions as far
back as Yosemite which was released in 2014, and
iOS 10 which was released in 2016. So we are safe to
remove this workaround which was added in 2013.

https://opensource.apple.com/source/xnu/xnu-1504.9.37/bsd/netinet/in.h
https://opensource.apple.com/source/xnu/xnu-1699.22.73/bsd/netinet/in.h
https://codereview.chromium.org/99923004/diff/200001/net/udp/udp_socket_libevent.cc

Change-Id: Ib67ec88aa618b977fe1f64a4862edda8c8ffefe1

Bug: 1128459
Change-Id: Ib67ec88aa618b977fe1f64a4862edda8c8ffefe1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2415193
Commit-Queue: Renjie Tang <renjietang@chromium.org>
Reviewed-by: David Schinazi <dschinazi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#808015}
GitOrigin-RevId: 298fd164249ca22d9c7567857e3809b3a4f36868